refactor: move encode_params from test to oauth2.py
